recipe/meta.yaml: fix Python/Meson specs
First, make Meson a build-time dependency.

Second, we had a filter to only build on Python 3.6 because I thought that
Python was only a build-time dependency. But that's not correct: this package
provides a couple of developer support tools written in Python. So for the
time being we need to include Python as a runtime dependency too, and I feel
like the safest approach is to matrix across all versions.

Third, remove the conda inspect commands.