`glib` needs `packaging` at runtime, mostly
As mentioned in the comment and associated pull request, it turns out
that in some scenarios the `gdbus-codegen` script is unavoidably
executed by a Python interpreter other than the one installed into the
"run" environment. So while this dep should help in most cases, there
might be times that you need to add `packaging` as a dep elsewhere.

(Part of the reason is that this script uses `/usr/bin/env python` as
its shebang line, but even if we rewrite that, Conda will re-rewrite it
back to use `/usr/bin/env` if the host prefix is very long ... which is
exactly the scenario encountered when this package is a build dependency
of another package build.)